extract to "dev-logger"
